added new AFNetworking logger
