+++ AFNetworkActivityConsoleLogger.m
@@ -1,0 +1,90 @@
+// AFNetworkActivityConsoleLogger.h
+//
+// Copyright (c) 2015 AFNetworking (http://afnetworking.com/)
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+#import "AFNetworkActivityConsoleLogger.h"
+
+@implementation AFNetworkActivityConsoleLogger
+
+- (id)init {
+    self = [super init];
+    if (!self) {
+        return nil;
+    }
+
+    self.level = AFLoggerLevelDebug;
+
+    return self;
+}
+
+
+- (void)URLSessionTaskDidStart:(NSURLSessionTask *)task {
+    NSURLRequest *request = task.originalRequest;
+
+    NSString *body = nil;
+    if ([request HTTPBody]) {
+        body = [[NSString alloc] initWithData:[request HTTPBody] encoding:NSUTF8StringEncoding];
+    }
+
+    switch (self.level) {
+        case AFLoggerLevelDebug:
+            NSLog(@"%@ '%@': %@ %@", [request HTTPMethod], [[request URL] absoluteString], [request allHTTPHeaderFields], body);
+            break;
+        case AFLoggerLevelInfo:
+            NSLog(@"%@ '%@'", [request HTTPMethod], [[request URL] absoluteString]);
+            break;
+        default:
+            break;
+    }
+}
+
+- (void)URLSessionTaskDidFinish:(NSURLSessionTask *)task withResponseObject:(id)responseObject inElapsedTime:(NSTimeInterval )elapsedTime withError:(NSError *)error {
+    NSUInteger responseStatusCode = 0;
+    NSDictionary *responseHeaderFields = nil;
+    if ([task.response isKindOfClass:[NSHTTPURLResponse class]]) {
+        responseStatusCode = (NSUInteger)[(NSHTTPURLResponse *)task.response statusCode];
+        responseHeaderFields = [(NSHTTPURLResponse *)task.response allHeaderFields];
+    }
+
+    if (error) {
+        switch (self.level) {
+            case AFLoggerLevelDebug:
+            case AFLoggerLevelInfo:
+            case AFLoggerLevelError:
+                NSLog(@"[Error] %@ '%@' (%ld) [%.04f s]: %@", [task.originalRequest HTTPMethod], [[task.response URL] absoluteString], (long)responseStatusCode, elapsedTime, error);
+            default:
+                break;
+        }
+    } else {
+        switch (self.level) {
+            case AFLoggerLevelDebug:
+                NSLog(@"%ld '%@' [%.04f s]: %@ %@", (long)responseStatusCode, [[task.response URL] absoluteString], elapsedTime, responseHeaderFields, responseObject);
+                break;
+            case AFLoggerLevelInfo:
+                NSLog(@"%ld '%@' [%.04f s]", (long)responseStatusCode, [[task.response URL] absoluteString], elapsedTime);
+                break;
+            default:
+                break;
+        }
+    }
+}
+
+@end
