+++ AFNetworkActivityLogger.h
@@ -1,6 +1,6 @@
 // AFNetworkActivityLogger.h
 //
-// Copyright (c) 2013 AFNetworking (http://afnetworking.com/)
+// Copyright (c) 2015 AFNetworking (http://afnetworking.com/)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -21,17 +21,7 @@
 // THE SOFTWARE.
 
 #import <Foundation/Foundation.h>
-
-typedef NS_ENUM(NSUInteger, AFHTTPRequestLoggerLevel) {
-  AFLoggerLevelOff,
-  AFLoggerLevelDebug,
-  AFLoggerLevelInfo,
-  AFLoggerLevelWarn,
-  AFLoggerLevelError,
-  AFLoggerLevelFatal = AFLoggerLevelOff,
-};
-
-typedef void(^AFLogger)(NSString *format);
+#import "AFNetworkActivityLoggerProtocol.h"
 
 /**
  `AFNetworkActivityLogger` logs requests and responses made by AFNetworking, with an adjustable level of detail.
@@ -42,21 +32,15 @@ typedef void(^AFLogger)(NSString *format);
  
  `AFNetworkActivityLogger` listens for `AFNetworkingOperationDidStartNotification` and `AFNetworkingOperationDidFinishNotification` notifications, which are posted by AFNetworking as request operations are started and finish. For further customization of logging output, users are encouraged to implement desired functionality by listening for these notifications.
  */
+
+typedef void(^AFLogger)(NSString *format);
+
 @interface AFNetworkActivityLogger : NSObject
 
 /**
- The level of logging detail. See "Logging Levels" for possible values. `AFLoggerLevelInfo` by default.
+ The set of loggers current managed by the shared activity logger. By default, this includes one `AFNetworkActivityConsoleLogger`
  */
-@property (nonatomic, assign) AFHTTPRequestLoggerLevel level;
-
-/**
- Only log requests conforming to the specified predicate, if specified. `nil` by default.
- 
- @discussion Each notification has an associated `NSURLRequest`. To filter out request and response logging, such as all network activity made to a particular domain, this predicate can be set to match against the appropriate URL string pattern.
- */
-@property (nonatomic, strong) NSPredicate *filterPredicate;
-
-@property (nonatomic, strong) AFLogger loggerBlock;
+@property (nonatomic, strong, readonly) NSSet <AFNetworkActivityLoggerProtocol> *loggers;
 
 /**
  Returns the shared logger instance.
@@ -64,50 +48,26 @@ typedef void(^AFLogger)(NSString *format);
 + (instancetype)sharedLogger;
 
 /**
- Start logging requests and responses.
+ Start logging requests and responses to all managed loggers.
  */
 - (void)startLogging;
 
 /**
- Stop logging requests and responses.
+ Stop logging requests and responses to all managed loggers.
  */
 - (void)stopLogging;
 
-@end
+@property (nonatomic, strong) AFLogger loggerBlock;
 
-///----------------
-/// @name Constants
-///----------------
 
 /**
- ## Logging Levels
+ Adds the given logger to be managed to the `loggers` set.
+ */
+- (void)addLogger:(id <AFNetworkActivityLoggerProtocol>)logger;
 
- The following constants specify the available logging levels for `AFNetworkActivityLogger`:
+/**
+ Removes the given logger from the `loggers` set.
+ */
+- (void)removeLogger:(id <AFNetworkActivityLoggerProtocol>)logger;
 
- enum {
- AFLoggerLevelOff,
- AFLoggerLevelDebug,
- AFLoggerLevelInfo,
- AFLoggerLevelWarn,
- AFLoggerLevelError,
- AFLoggerLevelFatal = AFLoggerLevelOff,
- }
-
- `AFLoggerLevelOff`
- Do not log requests or responses.
-
- `AFLoggerLevelDebug`
- Logs HTTP method, URL, header fields, & request body for requests, and status code, URL, header fields, response string, & elapsed time for responses.
- 
- `AFLoggerLevelInfo`
- Logs HTTP method & URL for requests, and status code, URL, & elapsed time for responses.
-
- `AFLoggerLevelWarn`
- Logs HTTP method & URL for requests, and status code, URL, & elapsed time for responses, but only for failed requests.
- 
- `AFLoggerLevelError`
- Equivalent to `AFLoggerLevelWarn`
-
- `AFLoggerLevelFatal`
- Equivalent to `AFLoggerLevelOff`
-*/
+@end
