+++ AFNetworkActivityLogger.m
@@ -1,6 +1,6 @@
 // AFNetworkActivityLogger.h
 //
-// Copyright (c) 2013 AFNetworking (http://afnetworking.com/)
+// Copyright (c) 2015 AFNetworking (http://afnetworking.com/)
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -21,21 +21,25 @@
 // THE SOFTWARE.
 
 #import "AFNetworkActivityLogger.h"
-#import "AFURLConnectionOperation.h"
 #import "AFURLSessionManager.h"
-
+#import "AFNetworkActivityConsoleLogger.h"
 #import <objc/runtime.h>
 
-static NSURLRequest * AFNetworkRequestFromNotification(NSNotification *notification) {
-    NSURLRequest *request = nil;
-    if ([[notification object] isKindOfClass:[AFURLConnectionOperation class]]) {
-        request = [(AFURLConnectionOperation *)[notification object] request];
-    } else if ([[notification object] respondsToSelector:@selector(originalRequest)]) {
-        request = [[notification object] originalRequest];
+static NSError * AFNetworkErrorFromNotification(NSNotification *notification) {
+    NSError *error = nil;
+    if ([[notification object] isKindOfClass:[NSURLSessionTask class]]) {
+        error = [(NSURLSessionTask *)[notification object] error];
+        if (!error) {
+            error = notification.userInfo[AFNetworkingTaskDidCompleteErrorKey];
+        }
     }
-
-    return request;
+    return error;
 }
+
+@interface AFNetworkActivityLogger ()
+@property (nonatomic, strong) NSMutableSet *mutableLoggers;
+
+@end
 
 @implementation AFNetworkActivityLogger
 
@@ -56,40 +60,38 @@ static NSURLRequest * AFNetworkRequestFromNotification(NSNotification *notificat
         return nil;
     }
 
-    self.level = AFLoggerLevelInfo;
+    self.mutableLoggers = [NSMutableSet set];
+
+    AFNetworkActivityConsoleLogger *consoleLogger = [AFNetworkActivityConsoleLogger new];
+    [self addLogger:consoleLogger];
 
     return self;
+}
+
+- (NSSet *)loggers {
+    return self.mutableLoggers;
 }
 
 - (void)dealloc {
     [self stopLogging];
 }
 
+- (void)addLogger:(id<AFNetworkActivityLoggerProtocol>)logger {
+    [self.mutableLoggers addObject:logger];
+}
+
+- (void)removeLogger:(id<AFNetworkActivityLoggerProtocol>)logger {
+    [self.mutableLoggers removeObject:logger];
+}
+
 - (void)startLogging {
     [self stopLogging];
-
-    [[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(networkRequestDidStart:) name:AFNetworkingOperationDidStartNotification object:nil];
-    [[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(networkRequestDidFinish:) name:AFNetworkingOperationDidFinishNotification object:nil];
-
-#if (defined(__IPHONE_OS_VERSION_MAX_ALLOWED) && __IPHONE_OS_VERSION_MAX_ALLOWED >= 70000) || (defined(__MAC_OS_X_VERSION_MAX_ALLOWED) && __MAC_OS_X_VERSION_MAX_ALLOWED >= 1090)
     [[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(networkRequestDidStart:) name:AFNetworkingTaskDidResumeNotification object:nil];
     [[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(networkRequestDidFinish:) name:AFNetworkingTaskDidCompleteNotification object:nil];
-#endif
 }
 
 - (void)stopLogging {
     [[NSNotificationCenter defaultCenter] removeObserver:self];
-}
-
-- (void)aflog:(NSString *)format, ... {
-    va_list args;
-    va_start(args, format);
-    if (self.loggerBlock) {
-        self.loggerBlock([[NSString alloc] initWithFormat:format arguments:args]);
-    } else {
-        NSLogv(format, args);
-    }
-    va_end(args);
 }
 
 #pragma mark - NSNotification
@@ -97,83 +99,47 @@ static NSURLRequest * AFNetworkRequestFromNotification(NSNotification *notificat
 static void * AFNetworkRequestStartDate = &AFNetworkRequestStartDate;
 
 - (void)networkRequestDidStart:(NSNotification *)notification {
-    NSURLRequest *request = AFNetworkRequestFromNotification(notification);
+    NSURLSessionTask *task = [notification object];
+    NSURLRequest *request = task.originalRequest;
 
     if (!request) {
         return;
     }
 
-    if (request && self.filterPredicate && [self.filterPredicate evaluateWithObject:request]) {
-        return;
-    }
-
     objc_setAssociatedObject(notification.object, AFNetworkRequestStartDate, [NSDate date], OBJC_ASSOCIATION_RETAIN_NONATOMIC);
 
-    NSString *body = nil;
-    if ([request HTTPBody]) {
-        body = [[NSString alloc] initWithData:[request HTTPBody] encoding:NSUTF8StringEncoding];
-    }
+    for (id <AFNetworkActivityLoggerProtocol> logger in self.loggers) {
+        if (request && logger.filterPredicate && [logger.filterPredicate evaluateWithObject:request]) {
+            return;
+        }
 
-    switch (self.level) {
-        case AFLoggerLevelDebug:
-            [self aflog:@"%@ '%@': %@ %@", [request HTTPMethod], [[request URL] absoluteString], [request allHTTPHeaderFields], body];
-            break;
-        case AFLoggerLevelInfo:
-            [self aflog:@"%@ '%@'", [request HTTPMethod], [[request URL] absoluteString]];
-            break;
-        default:
-            break;
+        [logger URLSessionTaskDidStart:task];
     }
 }
 
 - (void)networkRequestDidFinish:(NSNotification *)notification {
-    NSURLRequest *request = AFNetworkRequestFromNotification(notification);
-    NSURLResponse *response = [notification.object response];
-    NSError *error = [notification.object error];
+    NSURLSessionTask *task = [notification object];
+    NSURLRequest *request = task.originalRequest;
+    NSURLResponse *response = task.response;
+    NSError *error = AFNetworkErrorFromNotification(notification);
 
     if (!request && !response) {
         return;
     }
 
-    if (request && self.filterPredicate && [self.filterPredicate evaluateWithObject:request]) {
-        return;
-    }
-
-    NSUInteger responseStatusCode = 0;
-    NSDictionary *responseHeaderFields = nil;
-    if ([response isKindOfClass:[NSHTTPURLResponse class]]) {
-        responseStatusCode = [(NSHTTPURLResponse *)response statusCode];
-        responseHeaderFields = [(NSHTTPURLResponse *)response allHeaderFields];
-    }
-
-    NSString *responseString = nil;
-    if ([[notification object] respondsToSelector:@selector(responseString)]) {
-        responseString = [[notification object] responseString];
+    id responseObject = nil;
+    if (notification.userInfo) {
+        responseObject = notification.userInfo[AFNetworkingTaskDidCompleteSerializedResponseKey];
     }
 
     NSTimeInterval elapsedTime = [[NSDate date] timeIntervalSinceDate:objc_getAssociatedObject(notification.object, AFNetworkRequestStartDate)];
 
-    if (error) {
-        switch (self.level) {
-            case AFLoggerLevelDebug:
-            case AFLoggerLevelInfo:
-            case AFLoggerLevelWarn:
-            case AFLoggerLevelError:
-                [self aflog:@"[Error] %@ '%@' (%ld) [%.04f s]: %@", [request HTTPMethod], [[response URL] absoluteString], (long)responseStatusCode, elapsedTime, error];
-            default:
-                break;
+    for (id <AFNetworkActivityLoggerProtocol> logger in self.loggers) {
+        if (request && logger.filterPredicate && [logger.filterPredicate evaluateWithObject:request]) {
+            return;
         }
-    } else {
-        switch (self.level) {
-            case AFLoggerLevelDebug:
-                [self aflog:@"%ld '%@' [%.04f s]: %@ %@", (long)responseStatusCode, [[response URL] absoluteString], elapsedTime, responseHeaderFields, responseString];
-                break;
-            case AFLoggerLevelInfo:
-                [self aflog:@"%ld '%@' [%.04f s]", (long)responseStatusCode, [[response URL] absoluteString], elapsedTime];
-                break;
-            default:
-                break;
-        }
+
+        [logger URLSessionTaskDidFinish:task withResponseObject:responseObject inElapsedTime:elapsedTime withError:error];
     }
 }
